% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speechInput.R
\name{speechInput}
\alias{speechInput}
\title{Create a speech recognition input control}
\usage{
speechInput(inputId, command = "hey shiny *message")
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{command}{The voice command that will activate the speechInput. The
voice command must be exactly matched in order to send a value to the
speechInput. To capture particular parts of the command, there are two
symbols: "*" and ":". "*" captures any number of words, meanwhile ":"
captures exactly one word. Moreover, to make words optional, they should be
put between "(" and ")". \cr\cr
For example:
\itemize{
  \item "hello world" will trigger the input$inputId event, with no
        particular value, when it is said "hello world".
  \item "hello (my beatiful) world" will trigger the input$inputId event,
        with no particular value, when it is said "hello world" or "hello
        (my beatiful) world".
  \item "hello my *world" will trigger the input$inputId event, with all
        the words said after "hello my".
  \item "hello my :world" will trigger the input$inputId event, with the
        first word said after "hello my".
}}
}
\description{
Create an input control for entry of speech recognition text values
}
