% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopHeyshiny.R
\name{stopHeyshiny}
\alias{stopHeyshiny}
\title{Stop the SpeechRecognition}
\usage{
stopHeyshiny(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{Session object to send notification to.}
}
\description{
Stop the browser's SpeechRecognition task. This is useful to be called when
the Shiny app is going to be exited. If not, the browser's tab will keep
trying to perfom speech recognition.
}
