% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useHeyshiny.R
\name{useHeyshiny}
\alias{useHeyshiny}
\title{Set up a Shiny app to use heyshiny}
\usage{
useHeyshiny(language = "en")
}
\arguments{
\item{language}{The language the user will speak in. Check posibilites at:
https://github.com/TalAter/annyang/blob/master/docs/FAQ.md#what-languages-are-supported}
}
\description{
This function must be called from a Shiny app's UI in order for all other
heyshiny functions to work. \cr\cr
You can call useHeyshiny() from anywhere inside the UI, as long as the final
app UI contains the result of useHeyshiny().
}
